requirements.txt
apache-airflow==2.1.3
apache-airflow-providers-postgres==2.1.0
apache-airflow-providers-mysql==2.1.0
apache-airflow-providers-elasticsearch==2.0.2
apache-airflow-providers-ssh==2.1.0
apache-airflow-providers-sftp==2.1.0
apache-airflow[apache.atlas]
apache-airflow[cncf.kubernetes]
#apache-airflow-providers-sendgrid
SQLAlchemy         #==1.4.20
marshmallow-sqlalchemy             #==0.26.1
mysql-connector-python==8.0.22
mysqlclient==2.0.3
flask               #==2.0.1
itsdangerous              #==2.0.1
Flask-AppBuilder==3.3.2
Flask-WTF              #==0.15.1
Flask-caching==1.10.1
gunicorn==20.1.0
gevent
celery==5.1.2
flower==1.0.0
redis==3.5.3
setproctitle==1.2
python-daemon
blinker
httpx
elasticsearch==7.6.0
elasticsearch_dsl
kubernetes==11.0.0
waitress==2.0.0
astmonkey
#doltpy==2.0.9
#attrs==20.1.0
gitdb
gitdb-speedups
GitPython
requests
paramiko
sshtunnel
atlasclient==1.0.0
SpiffWorkflow


